backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panta-rhei-backend-api"
version = "1.0.0"
description = "Ocean Climate Data Backend API for Panta-rhei Web Globe"
requires-python = ">=3.10"
dependencies = [
    "copernicusmarine>=1.3.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["clients", "config", "discovery", "experiments", "utils"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]